CODEREVGENAI/backend/requirements-dev.txt
# Development Dependencies
# Includes all production dependencies plus development tools

-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.0

# Code Quality & Formatting
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# API Documentation
fastapi-openapi-docs==0.1.0

# Development Server
watchfiles==0.20.0

# Database Tools
mysql-connector-python==8.2.0

# Debugging
ipython==8.18.1
ipdb==0.13.13

# Environment Management
python-dotenv==1.0.0
